Input Manager now manages mouse state as well. Fixed bug in the TaskManager where it would hang indefinitely on shutdown. Main window redesigned to be just the screen area. The application can now be closed via Alt+F4 instead of only being able to close it via VS. The application is now also a Windows application without a console window attached to it.

+++ blowbox_console/console/console.cc
@@ -143,7 +143,7 @@ namespace console
 		network::ConsoleMessageTextLog msg;
 
 		memcpy(msg.log, log.c_str(), log.size());
-		msg.log_length = log.size();
+		msg.log_length = static_cast<short>(log.size());
 		msg.foregroundR = 150;
 		msg.foregroundG = 150;
 		msg.foregroundB = 150;
